Add files via upload
This was assignment 2 which had us creating a simple disk operating system. It taught me how files are stored and created, as well as how the computer "sees" the data because we had to show the directory entries as well.
I completed most of this assignment, except I didn't complete most of the functionality for the nested directory. I ran out of time for the submission but the main thing was because I started my solution without the nested functions in mind. So when I had to do those, I couldn't just use the implementations of the functions I had when I wasn't working with the nested files.